requirements.txt
Pillow
pycairo
tqdm
numpy
scipy
matplotlib

